Remove role update option from doctor information menu
Eliminate the options for updating a doctor's role and status in the menu. The display for "3. Update Role" and "4. Update Status" has been removed, retaining only the option to update status as "3. Update Status". Corresponding case statements for these updates have also been removed from the switch statement, disabling the functionality to update a doctor's role.